64 bits platform port
git-svn-id: https://svn.code.sf.net/p/auber/code/branches/tulip_3_0_0_dev@1105 f5d4c5f9-c943-4916-8a57-bd8605bf961c

+++ library/tulip/include/tulip/BiconnectedTest.h
@@ -53,7 +53,7 @@ private:
   void destroy(Graph *);
   BiconnectedTest(); //to ensure singleton
   static BiconnectedTest * instance;
-  stdext::hash_map<unsigned int,bool> resultsBuffer;
+  stdext::hash_map<unsigned long,bool> resultsBuffer;
 };
 }
 /*@}*/
